requirements.txt
requests==2.28.2
pandas==1.5.3
scikit-learn==1.2.2
joblib==1.2.0
python-dotenv==0.21.0
apache-airflow==2.5.0
dvc==2.24.1
